+++ lib/shiny_json_logic/operations/all.rb
@@ -8,10 +8,12 @@ module ShinyJsonLogic
     class All < Iterable::Base
       raise_on_dynamic_args!
 
-      def self.on_after(results, _scope_stack)
-        return false if results.empty?
+      def self.on_each(_item, filter, scope_stack)
+        throw(:early_return, false) unless Truthy.call(Engine.call(filter, scope_stack))
+      end
 
-        results.all? { |res| Truthy.call(res) }
+      def self.on_after(results, _scope_stack)
+        results.empty? ? false : true
       end
     end
   end